Added "Startup delay" to "Run applications" profile preferences. (4)

+++ phoneProfiles/src/main/java/sk/henrichg/phoneprofiles/ApplicationsCache.java
@@ -310,7 +310,7 @@ class ApplicationsCache {
             return 0;
     }
 
-    static long getStartApplicationDelay(String value) {
+    static int getStartApplicationDelay(String value) {
         if (value.length() > 2) {
             int startApplicationDelay = 0;
             String[] splits2 = value.split("/");
